REF: Move intersection calc to separate function
For art, move the intersection and distance calculation to a separate
function in order to reduce redundant calculations and remove redundant
code with art_fly_rotation. By computing the lengths and indices of
intersections and storing the result, iterative reconstruction is
immensely faster because for every iteration, these values are used
repeatedly. It's a trade off of memory usage for computation, but in
this case, the computation is very expensive, so it's worth it.