REF: Move intersection calc to separate function
For art, move the intersection and distance calculation to a separatefunction in order to reduce redundant calculations and remove redundantcode with art_fly_rotation. By computing the lengths and indices ofintersections and storing the result, iterative reconstruction isimmensely faster because for every iteration, these values are usedrepeatedly. It's a trade off of memory usage for computation, but inthis case, the computation is very expensive, so it's worth it.