MAIN: Move variable declarations to assignment locations
Since C99, we don't have to declare all the variables at the top
of the function. Therefore, it's more maintainable and safer to
have the declarations with the assignments because (1) it prevents
using variables before they are assigned and (2) it's more
maintainable because you don't have to look in multiple places
to see what a variable is or change it. In terms of speed, an
optimizing compiler shouldn't show any difference.