BUG: Index correct slice in multislice recon
During a previous refactor, the memory footprint of the function was
reduced by only allocating simdata a slice at a time. However, this
caused the usage of calc_simdata to always use the 0th slice of recon.
This commit fixes that by moving the recon pointer to the beginning of
each slice.